# add optimize kernel to hashcat