@wvu-r7 2020-01-18
# Release Notes

This adds the `-O` option to run an optimized kernel when invoking hashcat from Metasploit.